Replace uint with uint256 in Rebased.onUnstake function

+++ StakingContract/Rebase.sol
@@ -1930,7 +1930,7 @@ pragma solidity ^0.8.20;
 
 interface Rebased {
     function onStake(address user, address token, uint256 quantity) external;
-    function onUnstake(address user, address token, uint quantity) external;
+    function onUnstake(address user, address token, uint256 quantity) external;
 }
 
 interface WETH {
